# Rohafya/openapi.yaml
openapi: 3.0.0
info:
  title: New Gnu API
  version: 1.0.0
  description: >-
    # 🚀 Get started here


    This template guides you through CRUD operations (GET, POST, PUT, DELETE),
    variables, and tests.


    ## 🔖 **How to use this template**


    #### **Step 1: Send requests**


    RESTful APIs allow you to perform CRUD operations using the POST, GET, PUT,
    and DELETE HTTP methods.


    This collection contains each of these
    [request](https://learning.postman.com/docs/sending-requests/requests/)
    types. Open each request and click "Send" to see what happens.


    #### **Step 2: View responses**


    Observe the response tab for status code (200 OK), response time, and size.


    #### **Step 3: Send new Body data**


    Update or add new data in "Body" in the POST request. Typically, Body data
    is also used in PUT request.


    ```

    {
        "name": "Add your name in the body"
    }

     ```

    #### **Step 4: Update the variable**


    Variables enable you to store and reuse values in Postman. We have created a
    [variable](https://learning.postman.com/docs/sending-requests/variables/)
    called `base_url` with the sample request
    [https://postman-api-learner.glitch.me](https://postman-api-learner.glitch.me).
    Replace it with your API endpoint to customize this collection.


    #### **Step 5: Add tests in the "Scripts" tab**


    Adding tests to your requests can help you confirm that your API is working
...


    <img
...


...


...
        
...

...
servers:
...
paths:
...
  /add:
    post:
      summary: Add Roles
      responses: {}
      tags:
        - Roles API
      description: |-
        AJOUT D'UN ROLE

        ``` python
        @roles.route('/add', methods=["POST"])
        @cross_origin(supports_credentials=True)
        # @require_any_permission(["administration.roles.add_role"])
        @roles_required(["Admin"])
        @jwt_required()
        def add_roles():
            data = request.get_json()
            name = data.get("name")
            permission_codes = data.get("permissions", [])

            if not name:
                return {"Message": "Le nom du rôle est obligatoire."}, 400

            role = db.session.execute(select(Role).filter_by(name=name)).scalar_one_or_none()
            if role:
                return {"Message": f"Le rôle {name} existe déjà."}, 400

            # Création du rôle
            role = Role(name=name)

            # Récupération des permissions demandées
            if permission_codes:
                permissions = db.session.execute(
                    select(Permissions).filter(Permissions.code.in_(permission_codes))
                ).scalars().all()
                role.permissions = permissions

            db.session.add(role)
            db.session.commit()

            return {
                "id": role.id,
                "name": role.name,
                "permissions": [p.code for p in role.permissions]
            }, 201

         ```
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
            example:
              name: Comptable
...
  /notifications/types/add:
    post:
      summary: Add
      responses: {}
      tags:
        - Notifications
        - Types
      security:
        - BearerAuth: []
      description: >-
        ``` python

        @not_type.route('add', methods=["POST"])

        @cross_origin(supports_credentials=True)

        @jwt_required()

        # @roles_required("Admin")

        @require_any_permission(["administration.notifications_types.add_notification_type"])

        def add_notification_type():
            if request.method == 'POST' :
                name = request.get_json()['name']
                description = request.get_json().get('description', '')

                existing_type = db.session.execute(select(Notifications_Type).filter_by(name=name)).scalar_one_or_none()
                if existing_type :
                    return {"Message ": f" Le type de notification {name} existe déjà. "}

                new_type = Notifications_Type(name=name, description=description)
                db.session.add(new_type)
                db.session.commit()

                return new_type.to_dict()



         ```
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
            example:
              name: ALERT
              description: ALERT Type Of All Notifications.
...
  /notifications/types/:
    get:
      summary: All
      responses: {}
      tags:
        - Notifications
        - Types
      security:
        - BearerAuth: []
      description: >-
        ``` python

        @not_type.route("/", methods=['GET'])

        @cross_origin(supports_credentials=True)

        @jwt_required()

        # @roles_required("Admin")

        @require_any_permission(["administration.notifications_types.all_notification_types"])

        def all_notification_types():

            types = db.session.execute(select(Notifications_Type)).scalars().all()
            types_list = [ntype.to_dict() for ntype in types]

            return jsonify(types_list)



         ```
...
  /notifications/add:
    post:
      summary: Add Pour Tous
      responses: {}
      tags:
        - Notifications
        - Notifications
      security:
        - BearerAuth: []
      description: >-
        ``` python

        @general_notification.route('add', methods=["POST"])

        @tryton.transaction()

        @cross_origin(supports_credentials=True)

        # @roles_required("Admin")

        @jwt_required()

        @require_any_permission(["administration.notifications_general.add_notification"])

        def add_notifications():
            """AJOUTER DE NOUVELLES NOTIFICATIONS MANUELLEMENT DANS LA TABLE NOTIFICATIONS"""
            if request.method == "POST":
                data = request.get_json()
                types = data.get('types', None)
                title = data['title']
                message = data['message']
                created_at = datetime.now()
                all_users = data.get('all_users', None)
                user_id = data.get('user_id', None)
                notification = Notifications(
                    types=types,
                    title=title,
                    message=message,
                    created_at=created_at,
                    all_users=all_users,
                    user_id=user_id if not all_users else None
                )
                etiquettes = []
                if data.get('etiquettes'):
                    for etiquette_name in data.get('etiquettes', []):
                        etiquette = db.session.execute(select(Etiquettes_notifications).filter_by(name=etiquette_name)).scalar_one_or_none()
                        if etiquette:
                            etiquettes.append(etiquette)
                notification.etiquettes = etiquettes
                db.session.add(notification)
                db.session.commit()
                return {"Notifications" : notification.to_dict()}

         ```
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                message:
                  type: string
                types:
                  type: integer
                etiquettes:
                  type: array
                  items:
                    type: string
                all_users:
                  type: boolean
            example:
              title: Bienvenue sur l'Application de la PDMD
              message: IL SAGIT EN EFFET DE TOUT ET DE RIEN
              types: 3
              etiquettes:
                - COMMENTAIRE
                - COMMUNICATION
              all_users: true
...
  /device-info:
    get:
      summary: device-info
      responses: {}
      description: |-
        ``` python
            @app.route("/device-info", methods=["POST"])
            def device_info():
                data = request.json
                ip = request.headers.get("X-Forwarded-For", request.remote_addr)

                # Mettre à jour la dernière activité
                last_activity = UserActivity.query.filter_by(
                    ip=ip
                ).order_by(UserActivity.id.desc()).first()
                if last_activity :
                    last_activity.device_info = data
                    db.session.commit()

                return {"status": "saved"}

         ```
      requestBody:
        content:
          application/json:
            schema:
              type: object
  /whoami:
    get:
      summary: Who i Am
      responses: {}
      security:
        - BearerAuth: []
      description: |-
        ``` python
            @app.route("/whoami", methods=["GET"])
            @jwt_required()
            def who_am_i():
                current_user = get_jwt_identity()
                user = db.session.execute(select(User).filter_by(id=current_user)).scalar_one_or_none()
                if user:
                    return jsonify(user.to_dict())
                else:
                    return jsonify({"messsage": "User not found"})

         ```
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
security:
  - BearerAuth: []